Fix an infinite loop on 64-bit platforms; make the size of max_import_coord to be as big as the size of the 'fixed' type. Bug 688724.
DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6837 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gxhintn.h
@@ -115,7 +115,7 @@ typedef struct t1_hinter_s
 {   fraction_matrix ctmf;
     fraction_matrix ctmi;
     unsigned int g2o_fraction_bits;
-    unsigned int max_import_coord;
+    unsigned long max_import_coord;
     int32_t g2o_fraction;
     t1_glyph_space_coord orig_gx, orig_gy; /* glyph origin in glyph space */
     t1_glyph_space_coord subglyph_orig_gx, subglyph_orig_gy; /* glyph origin in glyph space */
